+++ __tests__/components/DateModified.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import { axe, toHaveNoViolations } from 'jest-axe'
-import DateModified from './DateModified'
+import DateModified from '../../components/DateModified'
 
 expect.extend(toHaveNoViolations)
