+++ __tests__/components/Header.test.js
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import { axe, toHaveNoViolations } from 'jest-axe'
-import Header from './Header'
-import en from '../locales/en'
-import fr from '../locales/fr'
+import Header from '../../components/Header'
+import en from '../../locales/en'
+import fr from '../../locales/fr'
 
 import { useRouter } from 'next/router'
