+++ __tests__/components/Layout.test.js
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import { axe, toHaveNoViolations } from 'jest-axe'
 
-import Layout from './Layout'
+import Layout from '../../components/Layout'
 import { useRouter } from 'next/router'
 
 // mocks useRouter to be able to use component' router.asPath
